finagent/requirements.txt
phidata
python-dotenv
yfinance
nsetools
packaging
duckduckgo-search
fastapi
uvicorn
groq